+++ xskillscore/core/probabilistic.py
@@ -505,8 +505,9 @@ def rps(
         ``.mean(member_dim))``. If ``fair==True``, forecasts must be boolean
         (True,False) or binary (0, 1) containing dimension ``member_dim``.
     category_edges : array_like
-        Category bin edges used to compute the CDFs. Bins include the left most edge, \
-        but not the right.
+        Category bin edges used to compute the CDFs. Similar to np.histogram, \
+        all but the last (righthand-most) bin include the left edge and exclude \
+        the right edge. The last bin includes both edges.
     dim : str or list of str, optional
         Dimension over which to compute mean after computing ``rps``.
         Defaults to None implying averaging over all dimensions.
@@ -670,7 +671,7 @@ def discrimination(
     observations,
     forecasts,
     dim=None,
-    probability_bin_edges=np.linspace(0, 1 + 1e-8, 6),
+    probability_bin_edges=np.linspace(0, 1, 6),
 ):
     """Returns the data required to construct the discrimination diagram for an event;
        the histogram of forecasts likelihood when observations indicate an event has
@@ -688,9 +689,10 @@ def discrimination(
         Dimension(s) over which to compute the histograms
         Defaults to None meaning compute over all dimensions.
     probability_bin_edges : array_like, optional
-        Probability bin edges used to compute the histograms.
-        Bins include the left most edge, \
-        but not the right. Defaults to 6 equally spaced edges between 0 and 1+1e-8
+        Probability bin edges used to compute the histograms. Similar to np.histogram, \
+        all but the last (righthand-most) bin include the left edge and exclude the \
+        right edge. The last bin includes both edges. Defaults to 6 equally spaced \
+        edges between 0 and 1
 
     Returns
     -------
@@ -759,7 +761,7 @@ def reliability(
     observations,
     forecasts,
     dim=None,
-    probability_bin_edges=np.linspace(0, 1 + 1e-8, 6),
+    probability_bin_edges=np.linspace(0, 1, 6),
     keep_attrs=False,
 ):
     """Returns the data required to construct the reliability diagram for an event;
@@ -778,9 +780,10 @@ def reliability(
             Dimension(s) over which to compute the histograms
             Defaults to None meaning compute over all dimensions.
         probability_bin_edges : array_like, optional
-            Probability bin edges used to compute the reliability.
-            Bins include the left most edge, \
-            but not the right. Defaults to 6 equally spaced edges between 0 and 1+1e-8
+            Probability bin edges used to compute the reliability. Similar to np.histogram,
+            all but the last (righthand-most) bin include the left edge and exclude the \
+            right edge. The last bin includes both edges. Defaults to 6 equally spaced \
+            edges between 0 and 1
         keep_attrs : bool, optional
             If True, the attributes (attrs) will be copied from the first input to
             the new one.
@@ -833,9 +836,13 @@ def reliability(
             N = np.zeros_like(r)
 
         for i in range(len(bin_edges) - 1):
-            # Follow xhistogram: all bins are half-open
-            # https://github.com/xgcm/xhistogram/issues/18
-            f_in_bin = (f >= bin_edges[i]) & (f < bin_edges[i + 1])
+            # Follow xhistogram behaviour: all bins are half-open, except for the right-most bin
+            # which adds an epsilon to the right edge
+            # see https://github.com/xgcm/xhistogram/issues/18
+            if i == (len(bin_edges) - 2):
+                f_in_bin = (f >= bin_edges[i]) & (f < (bin_edges[i + 1] + 1e-8))
+            else:
+                f_in_bin = (f >= bin_edges[i]) & (f < bin_edges[i + 1])
             o_f_in_bin = o & f_in_bin
             N_f_in_bin = f_in_bin.sum(axis=-1)
             N_o_f_in_bin = o_f_in_bin.sum(axis=-1)
@@ -957,11 +964,11 @@ def roc(
         Labeled array(s) over which to apply the function.
         If ``bin_edges=='continuous'``, forecasts are probabilities.
     bin_edges : array_like, str, default='continuous'
-        Bin edges for categorising observations and forecasts.
-        Bins include the left most edge, but not the right. ``bin_edges`` will be
-        sorted in ascending order. If ``bin_edges=='continuous'``, calculate
-        ``bin_edges`` from forecasts, equal to
-        ``sklearn.metrics.roc_curve(f_boolean, o_prob)``.
+        Bin edges for categorising observations and forecasts. Similar to np.histogram, \
+        all but the last (righthand-most) bin include the left edge and exclude the \
+        right edge. The last bin includes both edges. ``bin_edges`` will be sorted in \
+        ascending order. If ``bin_edges=='continuous'``, calculate ``bin_edges`` from \
+        forecasts, equal to ``sklearn.metrics.roc_curve(f_boolean, o_prob)``.
     dim : str, list
         The dimension(s) over which to compute the contingency table
     drop_intermediate : bool, default=False