+++ xskillscore/tests/test_probabilistic.py
@@ -431,7 +431,11 @@ def test_reliability_values(o, f_prob):
         for lat in f_prob.lat:
             o_1d = o.sel(lon=lon, lat=lat) > 0.5
             f_1d = (f_prob.sel(lon=lon, lat=lat) > 0.5).mean("member")
-            actual = reliability(o_1d, f_1d)
+            # scipy bins are only left-edge inclusive and 1e-8 is added to the last bin, whereas
+            # xhistogram the rightmost edge of xhistogram bins is included - mimic scipy behaviour
+            actual = reliability(
+                o_1d, f_1d, probability_bin_edges=np.linspace(0, 1 + 1e-8, 6)
+            )
             expected, _ = calibration_curve(
                 o_1d, f_1d, normalize=False, n_bins=5, strategy="uniform"
             )
@@ -442,7 +446,10 @@ def test_reliability_values(o, f_prob):
 def test_reliability_perfect_values(o):
     """Test values for perfect forecast"""
     f_prob = xr.concat(10 * [o], dim="member")
-    actual = reliability(o > 0.5, (f_prob > 0.5).mean("member"))
+    actual = reliability(
+        o > 0.5,
+        (f_prob > 0.5).mean("member"),
+    )
     expected_true_samples = (o > 0.5).sum()
     expected_false_samples = (o <= 0.5).sum()
     assert np.allclose(actual[0], 0)